Move HTTP/2 transport into h2.cpp
